Dead code removal. Renamed files more appropriately.
