+++ frontend/src/StatDisplay.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { StatAttributes } from "./StatAttributes";
+import { StatRow } from "./StatRow";
 
 export const StatDisplay = ({ player }) => {
   const formattedUrl = `/players/${player.attributes.player.name
@@ -17,7 +17,7 @@ export const StatDisplay = ({ player }) => {
         </Link>
       </span>
 
-      <StatAttributes attributes={player.attributes} />
+      <StatRow attributes={player.attributes} />
     </>
   );
 };
