+++ frontend/src/StatRow.js
@@ -1,15 +1,14 @@
 import React from "react";
 
-export const StatAttributes = ({ attributes }) => {
+export const StatRow = ({ attributes }) => {
   const renderYear = () => {
-    // debugger;
     if (window.location.pathname.split("/")[1] === "players") {
       const year = attributes.year.toString();
       return (
         <span className="stats-year-data">
-          <p>{year.slice(0, 4)}</p>
+          <p>'{year.slice(2, 4)}</p>
           <p>-</p>
-          <p>{year.slice(4)}</p>
+          <p>'{year.slice(6)}</p>
         </span>
       );
     }
